+++ mitra_cli/src/main.rs
@@ -50,7 +50,7 @@ async fn main() {
         SubCommand::FetchActor(cmd) => cmd.execute(&config, db_client).await.unwrap(),
         SubCommand::FetchActivity(cmd) => cmd.execute(&config, db_client).await.unwrap(),
         SubCommand::ReadOutbox(cmd) => cmd.execute(&config, db_client).await.unwrap(),
-        SubCommand::FetchReplies(cmd) => cmd.execute(&config, db_client).await.unwrap(),
+        SubCommand::LoadReplies(cmd) => cmd.execute(&config, db_client).await.unwrap(),
         SubCommand::FetchObject(cmd) => cmd.execute(&config, db_client).await.unwrap(),
         SubCommand::LoadPortableObject(cmd) => cmd.execute(&config, db_client).await.unwrap(),
         SubCommand::DeleteUser(cmd) => cmd.execute(&config, db_client).await.unwrap(),